requirements.txt
absl-py==0.9.0
asgiref==3.2.7
bleach==1.5.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==2.2
django-log-request-id==1.5.0
enum34==1.1.10
google-auth==1.16.0
google-auth-oauthlib==0.4.1
grpcio==1.29.0
h5py==2.10.0
html5lib==0.9999999
idna==2.8
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
mtcnn==0.1.0
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.34
Pillow==5.4.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-resize-image==1.1.19
pytz==2018.9
PyYAML==5.2
ratelim==0.1.6
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.12.0
sqlparse==0.3.1
tensorboard==2.1.1
tensorflow==1.4.0
tensorflow-estimator==2.1.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
urllib3==1.24.3
Werkzeug==1.0.1
